Moved business logic out of properties module
