+++ e2e-bfd-test/src/test/java/gov/cms/ab2d/testjobs/EndToEndBfdTests.java
@@ -2,7 +2,6 @@ package gov.cms.ab2d.testjobs;
 
 import gov.cms.ab2d.common.util.AB2DPostgresqlContainer;
 import gov.cms.ab2d.bfd.client.BFDClient;
-import gov.cms.ab2d.properties.dto.PropertiesDTO;
 import gov.cms.ab2d.common.model.Contract;
 import gov.cms.ab2d.job.model.Job;
 import gov.cms.ab2d.job.model.JobOutput;
@@ -82,9 +81,9 @@ import java.util.stream.Stream;
 import static gov.cms.ab2d.fhir.BundleUtils.EOB;
 import static gov.cms.ab2d.fhir.FhirVersion.R4;
 import static gov.cms.ab2d.fhir.FhirVersion.STU3;
-import static gov.cms.ab2d.properties.util.Constants.PCP_CORE_POOL_SIZE;
-import static gov.cms.ab2d.properties.util.Constants.PCP_MAX_POOL_SIZE;
-import static gov.cms.ab2d.properties.util.Constants.PCP_SCALE_TO_MAX_TIME;
+import static gov.cms.ab2d.common.util.PropertyConstants.PCP_CORE_POOL_SIZE;
+import static gov.cms.ab2d.common.util.PropertyConstants.PCP_MAX_POOL_SIZE;
+import static gov.cms.ab2d.common.util.PropertyConstants.PCP_SCALE_TO_MAX_TIME;
 import static java.util.stream.Collectors.toList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
