+++ hpms/src/main/java/gov/cms/ab2d/hpms/quartz/HPMSIngestQuartzSetup.java
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Configuration;
 /*
  *  todo: Don't know how to make this nice feature work using with the PropertiesService.  Instead, just allow the timer
  *      to fire, invoking HPMSIngestJob which immediately returns if not engaged.
- *      @ConditionalOnProperty(name = Constants.HPMS_INGESTION_ENGAGEMENT, havingValue = "engaged", matchIfMissing = false)
+ *      @ConditionalOnProperty(name = PropertyConstants.HPMS_INGESTION_ENGAGEMENT, havingValue = "engaged", matchIfMissing = false)
  */
 @Configuration
 public class HPMSIngestQuartzSetup {
