+++ properties/src/main/java/gov/cms/ab2d/properties/model/DataType.java
@@ -1,0 +1,8 @@
+package gov.cms.ab2d.properties.model;
+
+public enum DataType {
+    BOOLEAN,
+    INTEGER,
+    STRING,
+    FLOAT
+}
