+++ properties/src/test/resources/application.properties
@@ -7,3 +7,5 @@ spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 
 spring.jpa.hibernate.ddl-auto=none
+spring.liquibase.enabled=false
+spring.liquibase.contexts=test
