+++ worker/src/main/java/gov/cms/ab2d/worker/quartz/CoveragePeriodQuartzJob.java
@@ -4,7 +4,7 @@ import gov.cms.ab2d.properties.dto.PropertiesDTO;
 import gov.cms.ab2d.common.service.FeatureEngagement;
 import gov.cms.ab2d.properties.service.PropertiesAPIService;
 import gov.cms.ab2d.eventlogger.LogManager;
-import gov.cms.ab2d.properties.util.Constants;
+import gov.cms.ab2d.common.util.PropertyConstants;
 import gov.cms.ab2d.worker.processor.coverage.CoverageDriver;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,8 +17,10 @@ import java.time.DayOfWeek;
 import java.time.OffsetDateTime;
 import java.util.List;
 
-import static gov.cms.ab2d.properties.util.Constants.COVERAGE_SEARCH_OVERRIDE;
+import static gov.cms.ab2d.common.util.PropertyConstants.COVERAGE_SEARCH_DISCOVERY;
+import static gov.cms.ab2d.common.util.PropertyConstants.COVERAGE_SEARCH_OVERRIDE;
 import static gov.cms.ab2d.common.util.DateUtil.AB2D_ZONE;
+import static gov.cms.ab2d.common.util.PropertyConstants.COVERAGE_SEARCH_QUEUEING;
 import static gov.cms.ab2d.eventlogger.Ab2dEnvironment.PRODUCTION;
 import static gov.cms.ab2d.eventlogger.Ab2dEnvironment.SANDBOX;
 import static gov.cms.ab2d.eventlogger.events.SlackEvents.COVERAGE_UPDATES_FAILED;
@@ -27,17 +29,17 @@ import static gov.cms.ab2d.eventlogger.events.SlackEvents.COVERAGE_UPDATES_FAILE
  * Periodically update enrollment cached in the database by pulling enrollment from BFD.
  *
  * Typically this runs every Tuesday at midnight eastern time; however, you can override the schedule if enrollment arrives
- * at an unexpected time by setting {@link Constants#COVERAGE_SEARCH_OVERRIDE}.
+ * at an unexpected time by setting {@link PropertyConstants#COVERAGE_SEARCH_OVERRIDE}.
  *
  * Outside of production this feature will be disabled in most environments. To configure to run weekly,
  * set the following properties in the database:
  *
- *      - {@link Constants#COVERAGE_SEARCH_DISCOVERY} for all active contracts, find if those contracts are missing coverage periods
+ *      - {@link PropertyConstants#COVERAGE_SEARCH_DISCOVERY} for all active contracts, find if those contracts are missing coverage periods
  *         and create those missing coverage periods
  *          and loaded for the first time
- *      - {@link Constants#COVERAGE_SEARCH_QUEUEING} find all coverage periods missing enrollment or needing
+ *      - {@link PropertyConstants#COVERAGE_SEARCH_QUEUEING} find all coverage periods missing enrollment or needing
  *          enrollment updated, trigger those updates
- *      - {@link Constants#COVERAGE_SEARCH_OVERRIDE} normally this job only runs once a week, set this property to
+ *      - {@link PropertyConstants#COVERAGE_SEARCH_OVERRIDE} normally this job only runs once a week, set this property to
  *          override that configuration and force an update to enrollment.
  *
  * This only needs to run as often as BFD receives updated enrollment.
@@ -58,7 +60,7 @@ public class CoveragePeriodQuartzJob extends QuartzJobBean {
 
         try {
 
-            String discoveryEngagement = propertiesApiService.getProperty(Constants.COVERAGE_SEARCH_DISCOVERY);
+            String discoveryEngagement = propertiesApiService.getProperty(COVERAGE_SEARCH_DISCOVERY);
             FeatureEngagement disvoeryState = FeatureEngagement.fromString(discoveryEngagement);
 
             if (disvoeryState == FeatureEngagement.IN_GEAR) {
@@ -66,7 +68,7 @@ public class CoveragePeriodQuartzJob extends QuartzJobBean {
                 driver.discoverCoveragePeriods();
             }
 
-            String queueEngagement = propertiesApiService.getProperty(Constants.COVERAGE_SEARCH_QUEUEING);
+            String queueEngagement = propertiesApiService.getProperty(COVERAGE_SEARCH_QUEUEING);
             FeatureEngagement queueState = FeatureEngagement.fromString(queueEngagement);
 
             if (queueState == FeatureEngagement.IN_GEAR) {
@@ -90,10 +92,7 @@ public class CoveragePeriodQuartzJob extends QuartzJobBean {
             // override forces reload for all enabled contracts for last three months.
             // We don't want to be doing that over and over again.
             if (override) {
-                PropertiesDTO overrideUpdate = new PropertiesDTO();
-                overrideUpdate.setKey(COVERAGE_SEARCH_OVERRIDE);
-                overrideUpdate.setValue("false");
-                propertiesApiService.updateProperties(List.of(overrideUpdate));
+                propertiesApiService.updateProperty(COVERAGE_SEARCH_OVERRIDE, "false");
             }
         }
     }
