+++ worker/src/test/java/gov/cms/ab2d/worker/quartz/CoveragePeriodQuartzJobTest.java
@@ -4,7 +4,6 @@ import gov.cms.ab2d.properties.service.PropertiesAPIService;
 import gov.cms.ab2d.eventlogger.LogManager;
 import gov.cms.ab2d.worker.processor.coverage.CoverageDriver;
 import gov.cms.ab2d.worker.processor.coverage.CoverageDriverStub;
-import gov.cms.ab2d.worker.processor.coverage.CoverageProcessor;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,10 +15,10 @@ import org.quartz.JobExecutionException;
 import java.time.DayOfWeek;
 import java.time.OffsetDateTime;
 
-import static gov.cms.ab2d.properties.util.Constants.COVERAGE_SEARCH_DISCOVERY;
-import static gov.cms.ab2d.properties.util.Constants.COVERAGE_SEARCH_OVERRIDE;
+import static gov.cms.ab2d.common.util.PropertyConstants.COVERAGE_SEARCH_DISCOVERY;
+import static gov.cms.ab2d.common.util.PropertyConstants.COVERAGE_SEARCH_OVERRIDE;
 import static gov.cms.ab2d.common.util.DateUtil.AB2D_ZONE;
-import static gov.cms.ab2d.properties.util.Constants.COVERAGE_SEARCH_QUEUEING;
+import static gov.cms.ab2d.common.util.PropertyConstants.COVERAGE_SEARCH_QUEUEING;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
